Move visible_topics out of FilteringHelper into a TopicsFilteringHelper
visible_topics is routing/subscriptions-specific (view-level, non-
destructive filtering of live topic collections), so it does not belong
in the generic FilteringHelper that just renders the search box. Give it
its own Pro helper included alongside FilteringHelper.
